// src/types/TebexDueOnlineCommands.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface TebexDueOnlineCommands {
  commands: CommandO[];
}

export interface CommandO {
  id:         number;
  command:    string;
  payment:    number;
  package:    number;
  conditions: ConditionsO;
}

export interface ConditionsO {
  delay: number;
  slots: number;
}