// src/types/TebexDueQueue.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface TebexDueQueue {
  meta:    Meta;
  players: Player[];
}

export interface Meta {
  execute_offline: boolean;
  next_check:      number;
  more:            boolean;
}

export interface Player {
  id:   number;
  name: string;
  uuid: string;
}